requirements.txt
opencv-python==4.10.0.84
PySimpleGUI==5.0.7
pillow==11.0.0
pytesseract==0.3.13
SpeechRecognition==3.11.0